Rbk 036: Finalizacion de la plataforma Restaurant Aurea

+++ auth-node/src/categoria/categoria.service.js
@@ -1,4 +1,25 @@
 import Categoria from './categoria.model.js';
+import Producto from '../producto/producto.model.js';
+
+export class CategoriaError extends Error {
+    constructor(code, message, status = 400) {
+        super(message);
+        this.code = code;
+        this.status = status;
+    }
+}
+
+const withProductCount = async (categorias) => {
+    const counts = await Producto.aggregate([
+        { $group: { _id: '$idCategoria', total: { $sum: 1 } } },
+    ]);
+    const byCategory = new Map(counts.map((item) => [item._id.toString(), item.total]));
+
+    return categorias.map((categoria) => ({
+        ...categoria.toObject(),
+        productos: byCategory.get(categoria._id.toString()) || 0,
+    }));
+};
 
 const crearCategoria = async (data) =>
 {
@@ -12,7 +33,8 @@ const crearCategoria = async (data) =>
 
 const obtenerCategorias = async () =>
 {
-    return await Categoria.find().sort({ nombre: 1 });
+    const categorias = await Categoria.find().sort({ nombre: 1 });
+    return withProductCount(categorias);
 };
 
 const obtenerCategoriaPorId = async (id) =>
@@ -37,6 +59,15 @@ const actualizarCategoria = async (id, data) =>
 
 const eliminarCategoria = async (id) =>
 {
+    const productosAsociados = await Producto.countDocuments({ idCategoria: id });
+    if (productosAsociados > 0) {
+        throw new CategoriaError(
+            'CATEGORY_IN_USE',
+            `La categoria tiene ${productosAsociados} producto(s) asociado(s)`,
+            409
+        );
+    }
+
     const categoria = await Categoria.findByIdAndDelete(id);
 
     if (!categoria)
